Window Split
Added the ability to open the quick note in a horizontal or vertical
split. Added new keymaps for opening the windows in a vertical or
horizontal split. Update readme and quick-note doc to show the new
commands and feature.
