+++ lua/quick-note/init.lua
@@ -1,12 +1,15 @@
 local M = {} -- Naming convention for module
-local floating_window = require("quick-note.floating-window")
+-- local create_window = require("quick-note.create-window")
+local create_window = require("quick-note.create-window")
 
 local defaults = {
 	file_path = "~/.config/nvim/scratch.txt", -- @type string
 	keymaps = { -- @type table<string, string>
 		open_floating_window = "<leader>no", -- @type string
+		open_vertical_split_window = "<leader>nvs", -- @type string
+		open_horizontal_split_window = "<leader>nhs", -- @type string
 		close_buffer = "q", -- @type string
-		close_floating_window = "<leader>nc", -- @type string
+		close_window = "<leader>nc", -- @type string
 	},
 }
 
@@ -22,14 +25,24 @@ function M.setup(opts)
 
 	M.keymaps = keymaps
 
-	-- Global keymap to open the quick note window
+	-- Global keymap to open the quick note floating window
 	vim.keymap.set("n", M.keymaps.open_floating_window, function()
-		floating_window.open_floating_window(file_path, M.keymaps)
+		create_window.open_floating_window(file_path, M.keymaps)
+	end, { noremap = true, silent = true })
+
+	-- Global keymap to open the quick note verical window
+	vim.keymap.set("n", M.keymaps.open_vertical_split_window, function()
+		create_window.open_split_window(file_path, M.keymaps, "vertical")
+	end, { noremap = true, silent = true })
+
+	-- Global keymap to open the quick note horizontal window
+	vim.keymap.set("n", M.keymaps.open_horizontal_split_window, function()
+		create_window.open_split_window(file_path, M.keymaps, "horizontal")
 	end, { noremap = true, silent = true })
 
 	-- Global keymap to close the quick note window
-	vim.keymap.set("n", M.keymaps.close_floating_window, function()
-		floating_window.close_floating_window()
+	vim.keymap.set("n", M.keymaps.close_window, function()
+		create_window.close_window()
 	end, { noremap = true, silent = true })
 end
